pyproject.toml
[tool.poetry]
name = "google_sheets"
version = "0.1.0"
description = ""
authors = ["ilderebezov <ilderebezov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
install = "^1.3.5"
oauth2client = "^4.1.3"
google-api-python-client = "^2.48.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.1"
SQLAlchemy = "^1.4.36"
psycopg2 = "^2.9.3"
loguru = "^0.6.0"
python-dotenv = "^0.20.0"
pycbrf = "^1.1.0"
dash = "^2.4.1"
pandas = "^1.4.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
